requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
aiofiles
pyyaml